before add main menu

+++ FlappyBird/FlappyBird/GameConstants.h
@@ -10,7 +10,7 @@ static const float kRightMostX = 30.0f;
 static const float kBackSpeed = -1.0f;
 */
 static const float kPlayerSpeed = 5.0f;
-static const float kGravity = -9.8;
+static const float kGravity = -9.8f;
 static const float kAddedImpulse = 15.0f;
 static const float kImpulseDecreaseSpeed = 10.0f;
 static const int kMaxBarriers = 10;
@@ -18,6 +18,6 @@ static const float kBarrierTimeSpawn = 2.0f;
 static const float kBarrierXStartOffset = 25.0f;
 static const float kObjLiveMaxDistance = 30.0f;
 
-static const float kGameWidth = 640.0f;
-static const float kGameHeight = 480.0f;
+static const int kGameWidth = 640;
+static const int kGameHeight = 480;
